added a src directory, moved the data and utils into it and repathed what was missing. TO FIX :
stat pills info for gear cards.

Color matching and uniform icon and display.

Tier +, ++, +++ to match game

removing duplicate info in weapons
