+++ components/DamagePill.tsx
@@ -1,4 +1,4 @@
-// app/components/ui/DamagePill.tsx
+
 import React from "react";
 import { View, Text, Image, StyleSheet, ImageSourcePropType } from "react-native";
