+++ components/DungeonFilterGrid.tsx
@@ -1,4 +1,4 @@
-// components/DungeonFilterGrid.tsx
+
 import React from "react";
 import { View, Text, Pressable, StyleSheet } from "react-native";
 import { MaterialCommunityIcons } from "@expo/vector-icons";
