+++ components/RarityPill.tsx
@@ -1,4 +1,4 @@
-// app/components/ui/RarityPill.tsx
+
 import React from "react";
 import { View, Text, StyleSheet } from "react-native";
