+++ src/data/amulets.ts
