+++ src/data/armour.ts
