+++ src/data/gearImages.ts
