+++ src/data/techItems.ts
