+++ src/data/weapons.ts
